requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
django-paypal==1.0.0
django-registration-redux==2.5
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
python-memcached==1.59
pytz==2018.9
requests==2.21.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
